+++ R/classes.R
@@ -335,15 +335,19 @@ summary.epid <- function(object, ...){
   summ$iterations <- max(object@iteration)
   summ$total_records <- length(object)
   summ$total_episodes <- length(object[object@case_nm == 0])
+  # browser()
   x <- object[order(-object@wind_nm)]
-  x <- x[x@case_nm != -1]
-  x <- decode(x@wind_nm[!duplicated(x@.Data)])
+  x <- x[!duplicated(x@.Data)]
+  x <- decode(x@wind_nm[x@case_nm != -1])
   x[x == "Case"] <- "Fixed"
   x[x == "Recurrence"] <- "Rolling"
   summ$episode_type <- dst_tab(x = x, order_by_label = c("Fixed", "Rolling"))
-  x <- object[order(object@wind_id$wind_id1)]
-  x <- x[x@wind_nm == 1 & !duplicated(x@wind_id$wind_id1)]
+
+  x <- object[object@wind_nm == 1]
+  x <- x[!duplicated(x@wind_id$wind_id1)]
+  x <- x[order(x@.Data)]
   x <- rle(x@.Data)
+
   summ$recurrence <- dst_tab(x = paste0(x$lengths[order(x$lengths)]))
   if(length(summ$recurrence$values) > 0){
     summ$recurrence$values <- paste0(summ$recurrence$values, " times")